added all functions to one file and added require statement to replace functions
