+++ cms/functions.php
@@ -125,16 +125,32 @@ function returnProjectId(array $array, int $recId){
 }
 
 /* Doc Block
- * Adds paragraph from array content provided by database.
+ * Returns article from array content provided by database.
  *
  * @param $array array associative array provided by database.
  *
  * @return string the content of the arrays within the arrays pulled from the database.
  */
-function insertParagraph(array $array): string{
-    $paragraph = '';
+function returnArticle(array $array): string{
+    $article = '';
+
     foreach ($array as $value) {
-        $paragraph .= '<p>'.$value['aboutMeArticle'].'</p>';
+        $article .= $value['aboutMeArticle']."\n"."\n";
     }
+    return $article;
+}
+
+/* Doc Block
+ * Returns single paragraph from array content provided by database.
+ *
+ * @param $array array associative array provided by database.
+ *
+ * @return string the content of the arrays within the arrays pulled from the database.
+ */
+function returnParagraph(array $array, int $id): string{
+    $index = $id - 1;
+    $paragraphArray = $array[$index];
+    $paragraph = '';
+    $paragraph .= $paragraphArray['aboutMeArticle'];
     return $paragraph;
 }
